refactor: deduplicate advanced gas fee modal logic
